DTO instead of string in signal

+++ lw10/ShapesApp/ShapesApp/Controllers/DocumentController/IDocumentController.h
@@ -17,6 +17,6 @@ public:
 	virtual void Redo() = 0;
 
 	virtual void AddShape(const std::string& description) = 0;
-	virtual void AddImageItem(const std::string& imagePath, int width = 0, int height = 0) = 0;
+	virtual void AddImageItem(const std::string& imagePath, double width = 0, double height = 0) = 0;
 	virtual void RemoveItemAtIndex(size_t index) = 0;
 };
